RavenDB-24090 Use /cmpxchg EP for delete
