--- src/Raven.Studio/wwwroot/App/views/database/documents/deleteCompareExchangeProgress.html
@@ -1,41 +1,0 @@
-<div class="modal-dialog prevent-close" role="document">
-    <div class="modal-content">
-        <div class="modal-header">
-            <button type="button" class="close" data-bind="click: close, visible: !spinners.deleting()"
-                    aria-hidden="true">
-                <i class="icon-cancel"></i>
-            </button>
-            <h4 class="modal-title">Delete Compare Exchange Values</h4>
-        </div>
-        <div class="modal-body">
-            <div class="modal-details">
-                <h3 class="text-center text-danger" data-bind="visible: operationFailed">
-                    Operation failed!
-                </h3>
-                
-                <div class="details-item">
-                    <div class="row">
-                        <div class="col-sm-6 text-right">
-                            Total Items Processed:
-                        </div>
-                        <div class="col-sm-6 details-value">
-                            <span data-bind="text: processed().toLocaleString()"></span> / <span
-                            data-bind="text: total().toLocaleString()"></span>
-                        </div>
-                    </div>
-                </div>
-
-                <div class="progress">
-                    <div class="progress-bar" role="progressbar"
-                         data-bind="style: {'width': percentage() + '%'}">
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="modal-footer">
-            <button type="button" class="btn btn-default" data-bind="click: close, visible: !spinners.deleting()">
-                Close
-            </button>
-        </div>
-    </div>
-</div>
